багофикс для корректного создания/редактирования маршрутов
git-svn-id: http://svn.datateh.ru/alfrescoecm/Business%20Platform/trunk@6607 697f8a5b-bb4b-4f4c-ab90-ff1ad96ece97

+++ lecm-workflow/repo/src/main/java/ru/it/lecm/workflow/policies/RouteRolicy.java
@@ -41,15 +41,17 @@ public class RouteRolicy implements OnUpdateNodePolicy, OnCreateChildAssociation
 
 	@Override
 	public void onUpdateNode(final NodeRef routeRef) {
-		boolean hasTempAspect = nodeService.hasAspect(routeRef, LecmWorkflowModel.ASPECT_TEMP);
-		List<ChildAssociationRef> children = nodeService.getChildAssocs(routeRef, LecmWorkflowModel.ASSOC_ROUTE_CONTAINS_WORKFLOW_ASSIGNEES_LIST, RegexQNamePattern.MATCH_ALL);
-		int assigneesCount = 0;
-		for (ChildAssociationRef child : children) {
-			NodeRef assigneesListRef = child.getChildRef();
-			assigneesCount += nodeService.getChildAssocs(assigneesListRef, LecmWorkflowModel.ASSOC_WORKFLOW_ASSIGNEES_LIST_CONTAINS_ASSIGNEE, RegexQNamePattern.MATCH_ALL).size();
-		}
-		if (hasTempAspect && assigneesCount > 0) {
-			nodeService.removeAspect(routeRef, LecmWorkflowModel.ASPECT_TEMP);
+		if (nodeService.exists(routeRef)) {
+			boolean hasTempAspect = nodeService.hasAspect(routeRef, LecmWorkflowModel.ASPECT_TEMP);
+			List<ChildAssociationRef> children = nodeService.getChildAssocs(routeRef, LecmWorkflowModel.ASSOC_ROUTE_CONTAINS_WORKFLOW_ASSIGNEES_LIST, RegexQNamePattern.MATCH_ALL);
+			int assigneesCount = 0;
+			for (ChildAssociationRef child : children) {
+				NodeRef assigneesListRef = child.getChildRef();
+				assigneesCount += nodeService.getChildAssocs(assigneesListRef, LecmWorkflowModel.ASSOC_WORKFLOW_ASSIGNEES_LIST_CONTAINS_ASSIGNEE, RegexQNamePattern.MATCH_ALL).size();
+			}
+			if (hasTempAspect && assigneesCount > 0) {
+				nodeService.removeAspect(routeRef, LecmWorkflowModel.ASPECT_TEMP);
+			}
 		}
 	}
